Unify implementation of idx_to_lvl functions
