+++ lang/syntax/src/ast/traits/occurs.rs
@@ -2,7 +2,7 @@ use std::rc::Rc;
 
 use crate::ast::exp::Exp;
 use crate::ast::{Idx, Lvl};
-use crate::ctx::{Context, LevelCtx};
+use crate::ctx::LevelCtx;
 
 pub trait Occurs {
     fn occurs(&self, ctx: &mut LevelCtx, lvl: Lvl) -> bool;
