+++ lang/syntax/src/ctx/def.rs
@@ -33,6 +33,31 @@ impl<T> GenericCtx<T> {
     pub fn iter(&self) -> impl Iterator<Item = &[T]> {
         self.bound.iter().map(|inner| &inner[..])
     }
+
+    pub fn idx_to_lvl(&self, idx: Idx) -> Lvl {
+        let fst = self.bound.len() - 1 - idx.fst;
+        let snd = self.bound[fst].len() - 1 - idx.snd;
+        Lvl { fst, snd }
+    }
+
+    pub fn lvl_to_idx(&self, lvl: Lvl) -> Idx {
+        let fst = self.bound.len() - 1 - lvl.fst;
+        let snd = self.bound[lvl.fst].len() - 1 - lvl.snd;
+        Idx { fst, snd }
+    }
+
+    pub fn var_to_lvl(&self, var: Var) -> Lvl {
+        match var {
+            Var::Lvl(lvl) => lvl,
+            Var::Idx(idx) => self.idx_to_lvl(idx),
+        }
+    }
+    pub fn var_to_idx(&self, var: Var) -> Idx {
+        match var {
+            Var::Lvl(lvl) => self.lvl_to_idx(lvl),
+            Var::Idx(idx) => idx,
+        }
+    }
 }
 
 impl<T> From<Vec<Vec<T>>> for GenericCtx<T> {
@@ -62,21 +87,6 @@ pub trait Context: Sized {
 
     /// Pop a binder from the most-recently pushed telescope
     fn pop_binder(&mut self, elem: Self::Elem);
-
-    fn idx_to_lvl(&self, idx: Idx) -> Lvl;
-    fn lvl_to_idx(&self, lvl: Lvl) -> Idx;
-    fn var_to_lvl(&self, var: Var) -> Lvl {
-        match var {
-            Var::Lvl(lvl) => lvl,
-            Var::Idx(idx) => self.idx_to_lvl(idx),
-        }
-    }
-    fn var_to_idx(&self, var: Var) -> Idx {
-        match var {
-            Var::Lvl(lvl) => self.lvl_to_idx(lvl),
-            Var::Idx(idx) => idx,
-        }
-    }
 }
 
 /// Interface to bind variables to anything that has a `Context`
