+++ lang/syntax/src/ctx/values.rs
@@ -77,18 +77,6 @@ impl Context for TypeCtx {
         self.bound.last_mut().expect(err).pop().expect(err);
         self.shift_at_lvl(0..1, self.bound.len() - 1, (0, -1));
     }
-
-    fn idx_to_lvl(&self, idx: Idx) -> Lvl {
-        let fst = self.bound.len() - 1 - idx.fst;
-        let snd = self.bound[fst].len() - 1 - idx.snd;
-        Lvl { fst, snd }
-    }
-
-    fn lvl_to_idx(&self, lvl: Lvl) -> Idx {
-        let fst = self.bound.len() - 1 - lvl.fst;
-        let snd = self.bound[lvl.fst].len() - 1 - lvl.snd;
-        Idx { fst, snd }
-    }
 }
 
 impl ContextElem<TypeCtx> for &Binder {
